# public/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ORCA DAM API",
    "description": "Digital Asset Management API for ORCA DAM. Provides endpoints for managing assets, tags, and chunked file uploads.\n\n## Authentication\n\nAll endpoints except `/api/assets/meta` require authentication via Bearer token (Laravel Sanctum or JWT).\n\nInclude the token in the Authorization header:\n```\nAuthorization: Bearer YOUR_TOKEN_HERE\n```\n\n## Rate Limiting\n\nChunked upload endpoints are rate-limited to 100 requests per minute.",
    "version": "1.0.0",
    "contact": {
      "name": "ORCA DAM"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/chunked-upload/chunk": {
      "post": {
        "tags": ["Chunked Upload"],
        "summary": "Upload chunk",
        "description": "Upload a single chunk of the file. Chunks should be 10MB each (except the last chunk). Rate limited to 100 requests per minute.",
        "operationId": "uploadChunk",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "session_token": {
                    "type": "string",
                    "description": "Session token from init response"
                  },
                  "chunk_number": {
                    "type": "integer",
                    "minimum": 1,
                    "description": "Chunk number (1-indexed)"
                  },
                  "chunk": {
                    "type": "string",
                    "format": "binary",
                    "description": "Chunk data (max 15MB)"
                  }
                },
                "required": ["session_token", "chunk_number", "chunk"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chunk uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChunkUploadResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Chunk upload failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Laravel Sanctum API token"
      }
    },
    "schemas": {
...
      "ChunkUploadResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Chunk uploaded successfully"
          },
          "part_number": {
            "type": "integer",
            "example": 1
          },
          "etag": {
            "type": "string",
            "description": "S3 ETag for the uploaded part"
          },
          "uploaded_chunks": {
            "type": "integer",
            "example": 5
          },
          "total_chunks": {
            "type": "integer",
            "example": 10
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Error message"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "The given data was invalid."
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "example": {
              "filename": ["The filename field is required."]
            }
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Unauthenticated."
                }
              }
            }
          }
        }
      },
...
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      }
    }
  }
}